Public terminate_batch, volatile status
Exposed a terminate_batch function to users, even though it shouldn't be necessary to use this.

Also set the status for each batch record to be volatile, as each worker
spins on this during a batch while another worker sets it during batch
termination. I'm not sure that this is actually necessary as long as the
local variable is declared as a volatile pointer in the batch scheduler
function.