Debugging info, batch termination protocol
Added some extra asserts and changed the way debug printing works.  More
importantly, changed the way workers decided when to return from
execute_batch. The original design was to use a batch counter, and
execute_batch would continue executing until that specific batch was
done, returning to the batch scheduler function. Unfortunately, I don't
have a good way to make sure a batch is actually in progress before
execute_batch is called.

Instead, I check if the batch lock is owned in each iteration of the
loop in execute_batch. If it's not owned, we return and check the item
status, possibly trying to start another batch. This sort-of breaks the
modularity that execute_batch is supposed to provide, but oh well. I'm
pretty sure the whole batcher counter thing is completely unnecessary
now (just checking the batch lock owner more often), but it's still good
to have for debugging purposes.