Removing the blist size field.
I removed the blist size field. For an individual blist, this doesn't
really matter. It does change the way the split function works,
however. For now I am actually calculating the size of the list at the
beginning of the split.

This is obviously not ideal. An alternative is to use
fetch_and_increment with a size field. Kunal explained how to do it
without the size, but the problem is the array. At the beginning of the
split, before doing the actual splitting, I need to allocate an array to
hold the new list pointers.

Actually, I don't really need an array. Instead I can just make a linked
list, since each tl_node has a next pointer, anyway. That's what I'll do
next.

Also put the split function into a new file. Logically, it's a separate
unit, since an individual will never need to split; it's only used in
the context of an upper-level structure that contains bottom lists.