Split using list instead of array.
Splitting a bottom list now produces a list of tl_nodes, each pointing
to a new bottom list, rather than producing an array. The original
tl_node is maintained, but now points to the beginning of this list
rather than a bottom list.

I've done this to get around needing a size value to allocate the array of new nodes.