Size field vs num_leaves
The 'size' field for a tl_node is now also a num_leaves field. This is
not really necessary, and really I should remove the original size field
and just call it num_leaves, rather than using a union. The new name
makes it clearer that this value does not count the internal nodes on
the way to the leaves.