Relabel uses list of new split nodes.
Rather than using an array of new split nodes for each split leaf, the
relabel (specifically, the build_array_of_leaves function) function now
uses the list provided by split, rather than an array.

Note that I do not patch the new lists into the list of leaves. I've
made a corresponding note of this in om.c. This is because it would be a
little extra work because naively doing this results in a race condition
between freeing the old tl_nodes and using the next/prev pointers. I'm
pretty sure we could get around this, but I don't see any reason to
because we're not using the list of leaves for any purpose other than
relabeling after splitting.