Changed a test case to reflect a slightly different relable implementations.
The test cases currently depend a lot on exactly how the rebalance
operations decides to split an array of leaves. Ideally, there will be
no (or few) such tests, and the tests would be more general, e.g. just
making sure none of the internal nodes are too heavy after a rebalance.