New public function to get batch id.
Exposed a new public function to return the current batch id of a
worker. This is used in the race detection tooling to make sure we don't
do anything while in a batch and so that the tool doesn't need to know
the internals of a worker's local state structure.