Interval timing bug
Fixed an interval timing bug -- I was storing all interval times as
integer milliseconds, converting each from seconds and
nanoseconds (struct timespec). The problem was that malloc calls are
typically very fast -- under 1 ms. So when I did integer division for
the conversion, the result was often 0. I've fixed this to represent the
time as a double now, performing the correct conversion.