Autoconf version required decreased
The required autoconf version was 2.64. I reduced it to 2.63. I don't
think 2.64 is actually required, and because Shakespeare is Scientific
Linux, it has really old packages. I can't even find a repo with
decent version of most packages! The versions of git, emacs, autoconf,
and many others are ancient, and I don't see way to update other than
to compile a bunch of packages (and their dependencies) from source,
which I don't have time for.