Merging Angelina's code.
Merged Angelina's code with mine. The shadow stack is not correctly
handled. Also, there is a race in my code when steal happens. However,
hopefully that is related to not handling the shadow stack
correctly. Also, I had to disable calls to anything in print_addr.cpp
as clang doesn't like the gcc c++11 headers on shakespeare.